api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dateutil==2.8.2
